+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
   - Wizard strength regen under 30% is now increased by by 1% per 100 spy resilience (from 1% per 250)
   - Wizard resilience is now only increased by Fireball
   - Peasant change due to wizard resilience is now increased by 10% of your total pop growth (Plague, Harmony, temples, techs, etc.)
+- Lightning repair bonus now scales based on total damage received, up to 100% at 10% damage received (from constant 100%)
 - Spy/Wizard ratio required for maximum damage reduction from black ops now scales up during the round: 0.5 at Day 4, 1.0 at Day 24, 1.5 at Day 44 (from 1.0 all round)
 
 ## [1.31.0] - 2022-09-02